feat(auth): add header based authentication
This makes it possible to use API Tokens as headers instead of having to set JWTs from some OAuth2
flow. I also removed the "PollinationAuth" method because it doesn't work now that we migrated to
Google Identity Platform. I will aim to add a nice OAuth2 style login for Pollination in the new
year (much like the gcloud API logs in users).